* mcc/Dispatcher.c: if a BetterString gadget moves into 'Inactive' state,   eventually selected content will be automatically unselected as it is done   with the default MUI string class. * misc: switched all OS4 builds to compile against newlib.library. * misc: updated 'mccheader.c' include file from latest TextEditor.mcc   sources and also applied the new Debug.c,h interface to all our subprojects. * mcc/hotkeystring/HandleInput.c: fixed the OS4 build to use the new media   key definitions instead.

+++ mcc/BetterString-Test.c
@@ -239,7 +239,7 @@ int main(void)
   		set(window, MUIA_Window_ActiveObject, bstring);
   		set(window, MUIA_Window_Open, TRUE);
 
-  		while(DoMethod(app, MUIM_Application_NewInput, &sigs) != MUIV_Application_ReturnID_Quit)
+  		while((LONG)DoMethod(app, MUIM_Application_NewInput, &sigs) != MUIV_Application_ReturnID_Quit)
   		{
   			if(sigs)
   			{
